Implemented Basic Drag-and-Drop
Created the TableTop scene, which will be the primary gameplay scene.
Added a temporary button on the MainMenu to go directly to the Tabletop scene, for convenient prototyping.

Created a Token scene. This is the Node that will represent any moveable asset that can be placed on the board, such as player characters or monsters. Implemented relatively simplistic click-and-drag functionality.

Created a GridCell scene. This is the Node that will be used to snap a Token into a specific place. Reacts to a Token being dragged into or out of its collision area.

Created some simple images to use in prototyping/testing the Token and GridCell nodes.