Set Up TileMap of GridCells
Reset position of the Area2D Node in the GridCell scene.

Added a TileMap, renamed Grid, in the TableTop scene and configured it to be able to place the GridCell scene as tiles.

Reduced the scale of the Token's CollisionShape to help a little with more accurately snapping to the correct GridCell. A true solution will require more complex logic but this will work for the moment.