Resized Images/GridCells
Created a new 80x80 box-transparent so the GridCell size could be based on a similar scale as typical resolutions.
Reduced the scale of the GridCell's CollisionShape2D, resolved most of the major snapping issues.
Increased fixed resolution for grid generation to 1080p.
Changed started y offset so the first row will align with the top of the viewport.