Basic Login Mock Up
Created SettingsMenu scene. The user will set their User Id here (this is not final, just for the mock up).

Created a SceneChange script to simplify creating buttons for navigating between scenes.

Removed the temporary "Go to Tabletop" button. Now when the user successfully "logs in," they are automatically moved to the TableTop scene. A successful login, for the purposes of this mock up, is simply passing a UserId which actually exists to the /login route of the server.

Updated DoConnect script to deal with the changes to vtt-server. Also, handle the responses it receives from the server more robustly.